aws_backend/requirements.txt
boto3==1.34.131
requests==2.31.0
PyJWT==2.8.0
stripe==5.5.0
# bcrypt==3.2.2  # Removed - using hashlib instead for Lambda compatibility
python-dotenv==1.0.0
assemblyai==0.24.0
openai==1.35.12
yt-dlp==2023.12.30
mutagen==1.47.0